Add some more censuses from Africa
