+++ public/data/census/axl/censusList.txt
@@ -1,14 +1,18 @@
+axl.bj
 axl.cd
 axl.cx
 axl.gl
 axl.km
+axl.ma
 axl.mv
 axl.nf
+axl.ng
 axl.no
 axl.pn
 axl.qa
 axl.sg
 axl.sl
+axl.ss
 axl.sy
 axl.tg
 axl.ye
