+++ public/data/census/data.un.org/et.tsv
@@ -61,7 +61,7 @@ zay	Zeysigna	18,512
 tsb	Tsamaigna	18,048
 cra	Charrigna	13,100
 dox	Mosiyagna	9,552
-kmw	Komigna	8,534
+xom	Komigna	8,534
 bcq/merr1238	Merigna	8,159
 kun	Kunamigna	7,428
 muz	Mursygna	7,386
