+++ public/data/census/official/bj2013.tsv
@@ -1,0 +1,75 @@
+#codeDisplay		bj2013
+#nameDisplay		Benin 2013
+#isoRegionCode	BJ	
+#yearCollected	2013	
+### Language Criteria ###		
+#languageUse	Speaks	
+#acquisitionOrder	L1	
+### Population Surveyed ###		
+#population	9787389	
+### Data Constraints ###		
+#quantity		
+### Creator ###		
+#collectorType	Government	
+#collectorName	Institut National de la Statistique et de la Démographie	
+#collectorNameShort	INStaD Benin	
+#presentedBy	CLEAR Global	
+### Source Document ###		
+#url	https://clearglobal.org/languages-of-benin-interactive-en/	
+#dateAccessed	2026-08-04	
+#sectionName	2.2 Les langues nationales	
+fon	Fon	2033740
+bba	Baatonum	871723
+ful/fue	Borgu Fulfulde	846296
+ajg	Aja	803664
+guw	Gun	628424
+nqg	Southern Nago	512328
+ayb	Ayizo Gbe	353618
+ddn	Dendi	318738
+mxl	Maxi Gbe	240616
+sxw	Saxwe Gbe	240480
+guw	Toli Gbe	235157
+wem	Weme Gbe	220505
+beh	Biali	190818
+tbz	Ditammari	189200
+pil	Yom	165717
+yor	Yoruba	137958
+idd	Ede Idaca	136939
+gej	Gen	134985
+dop	Lekpa	123428
+kqk	Kotafon Gbe	122739
+ijj	Ede Ije	120097
+tfi	Tofin Gbe	112419
+ife	Ifè	111130
+wwa	Waama	86817
+xwl	Western Xwla Gbe	85845
+bqc	Boo	80476
+fra	French	75842
+dje	Zarma	69318
+xwe	Xwela Gbe	59047
+gux	Gourmanchéma	51159
+wci	Waci Gbe	47782
+cbj	Ede Cabe	45501
+ntm	Nateni	41739
+hau	Hausa	41553
+mkl	Mokole	33006
+kbp	Kabiye	28481
+blo	Anii	26317
+gbh	Defi Gbe	26184
+kdh	Tem	21202
+guw	Seto Gbe	17872
+mql	Mbelime	13223
+gng	Gagamba	12218
+mul	Other	11569
+aug	Aguna	9983
+gng	Gando	9981
+soy	Mioybe	8836
+eng	English	6522
+mis	Contimba	6212
+ibo	Igbo	5205
+nqk	Partago	4222
+kdh	Kotokoli	3239
+ful/fuh	Gorgal Fulfulde	2865
+ewe	Ewe	2207
+pil/mis	Taneka	1176
+bqc	Boko	1070
