+++ public/data/census/unofficial/censusList.txt
@@ -19,3 +19,4 @@ ir
 is
 ke
 sn
+tz
