+++ public/data/tc/organizations.tsv
@@ -43,3 +43,5 @@ INE Angola	National Institute of Statistics	Instituto Nacional De Estatísticas	
 CBS Israel	Central Bureau of Statistics	הלשכה המרכזית לסטטיסטיקה	IL		https://www.cbs.gov.il/
 VNSO	Vanuatu National Statistics Office		VU		https://vnso.gov.vu/
 IOM Cabinet Office	Isle of Man Cabinet Office		IM		https://www.gov.im/
+INStaD Benin	Institut National de la Statistique et de la Démographie		BJ		https://nada.instad.bj/
+CLEAR Global	CLEAR Global		IE		https://clearglobal.org/
