+++ src/widgets/details/CensusDetails.tsx
@@ -1,16 +1,17 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 
-import HoverableObjectName from '@features/layers/hovercard/HoverableObjectName';
+import { useDataContext } from '@features/data/context/useDataContext';
+import EntityMap from '@features/map/EntityMap';
 import LocalParamsProvider from '@features/params/LocalParamsProvider';
+import { ObjectType } from '@features/params/PageParamTypes';
 
 import { CensusData } from '@entities/census/CensusTypes';
 
-import { toTitleCase } from '@shared/lib/stringUtils';
-import ExternalLink from '@shared/ui/ExternalLink';
-
 import TableOfLanguagesInCensus from '../tables/TableOfLanguagesInCensus';
 
-import DetailsField from './ui/DetailsField';
+import CensusPopulationCharacteristics from './sections/CensusPopulationCharacteristics';
+import CensusPrimarySection from './sections/CensusPrimarySection';
+import CensusSourceSection from './sections/CensusSourceSection';
 import DetailsSection from './ui/DetailsSection';
 
 type Props = {
@@ -18,6 +19,15 @@ type Props = {
 };
 
 const CensusDetails: React.FC<Props> = ({ census }) => {
+  const { getLanguage } = useDataContext();
+  const languages = useMemo(
+    () =>
+      Object.keys(census.languageEstimates)
+        .map((langID) => getLanguage(langID))
+        .filter((lang) => lang != null),
+    [census.languageEstimates, getLanguage],
+  );
+
   return (
     <div className="Details">
       <CensusPrimarySection census={census} />
@@ -28,160 +38,12 @@ const CensusDetails: React.FC<Props> = ({ census }) => {
           <TableOfLanguagesInCensus census={census} />
         </LocalParamsProvider>
       </DetailsSection>
+      <DetailsSection title="Languages on Map">
+        <LocalParamsProvider overrides={{ objectType: ObjectType.Language, limit: 1000 }}>
+          <EntityMap entities={languages} maxWidth={1000} />
+        </LocalParamsProvider>
+      </DetailsSection>
     </div>
-  );
-};
-
-function CensusPrimarySection({ census }: { census: CensusData }) {
-  const { territory, isoRegionCode, domain, proficiency, acquisitionOrder, languageUse } = census;
-  return (
-    <DetailsSection title="Primary Information">
-      <DetailsField title="Territory">
-        {territory != null ? (
-          <HoverableObjectName object={territory} />
-        ) : (
-          <span>{isoRegionCode}</span>
-        )}
-      </DetailsField>
-      <DetailsField title="Year">{census.yearCollected}</DetailsField>
-      {languageUse != null && <DetailsField title="Language Use">{languageUse}</DetailsField>}
-      {proficiency != null && <DetailsField title="Proficiency">{proficiency}</DetailsField>}
-      {acquisitionOrder != null && (
-        <DetailsField title="Acquisition Order">{acquisitionOrder}</DetailsField>
-      )}
-      {domain != null && <DetailsField title="Where language used">{domain}</DetailsField>}
-    </DetailsSection>
-  );
-}
-
-function CensusPopulationCharacteristics({ census }: { census: CensusData }) {
-  const {
-    age,
-    geographicScope,
-    languagesIncluded,
-    notes,
-    population,
-    populationSource,
-    populationSurveyed,
-    populationWithPositiveResponses,
-    quantity,
-    residenceBasis,
-    responsesPerIndividual,
-    sampleRate,
-  } = census;
-
-  return (
-    <DetailsSection title="Population Characteristics">
-      <DetailsField title="Overall population">{population.toLocaleString()}</DetailsField>
-      {populationSource && (
-        <DetailsField title="Source for overall population">
-          {populationSource.startsWith('http') ? (
-            <ExternalLink href={populationSource} />
-          ) : (
-            populationSource
-          )}
-        </DetailsField>
-      )}
-      {populationWithPositiveResponses && (
-        <DetailsField title="Responding Population">
-          {populationWithPositiveResponses.toLocaleString()}
-        </DetailsField>
-      )}
-      {populationSurveyed && (
-        <DetailsField title="Surveyed Population">
-          {populationSurveyed.toLocaleString()}
-        </DetailsField>
-      )}
-      {sampleRate ? (
-        <DetailsField title="Sample rate">
-          {typeof sampleRate === 'number' ? (sampleRate * 100).toLocaleString() + '%' : sampleRate}
-        </DetailsField>
-      ) : populationSurveyed ? (
-        <DetailsField title="Sample rate">
-          {((populationSurveyed / population) * 100).toLocaleString()}%
-        </DetailsField>
-      ) : null}
-      {languagesIncluded && (
-        <DetailsField title="Languages Included">{languagesIncluded}</DetailsField>
-      )}
-      {geographicScope && <DetailsField title="Geographic Scope">{geographicScope}</DetailsField>}
-      {residenceBasis && <DetailsField title="Residence Basis">{residenceBasis}</DetailsField>}
-      {age && <DetailsField title="Age">{age}</DetailsField>}
-      {responsesPerIndividual && (
-        <DetailsField title="Responses per Individual">{responsesPerIndividual}</DetailsField>
-      )}
-      {quantity && <DetailsField title="Quantity Provided">{toTitleCase(quantity)}</DetailsField>}
-      {notes && <DetailsField title="Notes">{notes}</DetailsField>}
-    </DetailsSection>
-  );
-}
-
-function CensusSourceSection({ census }: { census: CensusData }) {
-  const {
-    author,
-    citation,
-    columnName,
-    collectorType,
-    dateAccessed,
-    datePublished,
-    documentName,
-    presentedBy,
-    sectionName,
-    tableName,
-    url,
-  } = census;
-
-  return (
-    <DetailsSection title="Source">
-      <DetailsField title="Source type">{collectorType}</DetailsField>
-      <CensusCollectorNameDisplay census={census} />
-      {author && <DetailsField title="Author">{author}</DetailsField>}
-      {(census.presenter || presentedBy) && (
-        <DetailsField title="Presented by">
-          {census.presenter ? <HoverableObjectName object={census.presenter} /> : presentedBy}
-        </DetailsField>
-      )}
-      {url && (
-        <DetailsField title="URL">
-          <ExternalLink href={url} />
-        </DetailsField>
-      )}
-      {documentName && <DetailsField title="Document Name">{documentName}</DetailsField>}
-      {sectionName && <DetailsField title="Section Name">{sectionName}</DetailsField>}
-      {tableName && <DetailsField title="Table Name">{tableName}</DetailsField>}
-      {columnName && <DetailsField title="Column Name">{columnName}</DetailsField>}
-      {citation && <DetailsField title="Citation">{citation}</DetailsField>}
-      {datePublished && (
-        <DetailsField title="Date Published">
-          {new Date(datePublished).toLocaleDateString()}
-        </DetailsField>
-      )}
-      {dateAccessed && (
-        <DetailsField title="Date Accessed">
-          {new Date(dateAccessed).toLocaleDateString()}
-        </DetailsField>
-      )}
-    </DetailsSection>
-  );
-}
-
-const CensusCollectorNameDisplay: React.FC<{ census: CensusData }> = ({ census }) => {
-  const { collectorName, collectorNameShort, collector } = census;
-  if (!collectorName && !collectorNameShort && !collector) return null;
-
-  if (collector) {
-    return (
-      <DetailsField title="Collected by">
-        <HoverableObjectName object={collector} />
-      </DetailsField>
-    );
-  }
-
-  return (
-    <DetailsField title="Collected by">
-      {collectorName ? collectorName : collectorNameShort}
-      {collectorName && collectorNameShort && ` aka ${collectorNameShort}`}
-    </DetailsField>
   );
 };
